Renamed to checkFileExists
